Disk Pizza

+++ DiskPizza/DiskPizza.Models/Produto.cs
@@ -11,6 +11,6 @@ namespace DiskPizza.Models
         public int       Id      { get; set; }
         public string    Nome    { get; set; }
         public string    Tipo    { get; set; }
-        public float     Preco   { get; set; }
+        public decimal     Preco   { get; set; }
     }
 }
